extract metro config and package.json code
